requirements.txt
# ==========================
# requirements.txt (발췌)
# ==========================
pymongo
pandas
python-dotenv